Create the initial CaffeineCache implementation
This is the initial implementation of a `CaffeineCache` (caffeine is a highly performant in-memory cache) that will soon replace the `BasicCache`. The `BasicCache` was shown to be the limiter in several profiles and lacks a priority queue for items with expiry.

The new dependencies have been added and the code will compile, however, this is NOT REPLACING the `BasicCache` yet. More profiles and tweaks need to be done.