Upgrade anorm, postgresql, and postgis-jdbc dependencies
This commit updates the versions of the following dependencies in build.sbt:

- anorm and anorm-postgres from 2.6.10 to 2.7.0
- postgresql from 42.5.0 to 42.7.3
- postgis-jdbc from 2021.1.0 to 2023.1.0

The upgrade of these dependencies ensures that we are using the latest stable releases, which include various improvements and bug fixes. The URLs for the release notes of postgresql and postgis-jdbc have been added for easy reference.

This change is part of the ongoing effort to keep the project dependencies up-to-date.