Introduce maproulette.secret.key conf and MAPROULETTE_SECRET_KEY env
MapRoulette API had a dependency on the Play framework's secret key, and this patch separates those keys.
The MapRoulette secret is used to encrypt things specific to maproulette and it no longer depends on the Play application key.

Play 2.9 introduced a change where its application key must be 32 bytes or longer, and this impacted MapRoulette's cryptography with its own internal items. It's easier to manage these separately, similar to how the OSM secrets are not used to encrypt data stored within MapRoulette.