Remove quotes around HOCON substitutions
In HOCON, variable substitutions are not performed if the replacement part is in quotes.

From the manual:

Substitutions are not parsed inside quoted strings. To get a string containing a substitution, you must use value concatenation with the substitution in the unquoted portion:

key : ${animal.favorite} is my favorite animal